Adds generating images to run_mnist_exp.py.
Added a gen/mnist output directory, and modified the script to, when all
the training is completed, generate one batch-worth of images using the
now trained model.

This is dependent on b4602c9 and 8fca21e, which modify
infogan/algos/infogan_trainer.py to allow reusing the trained variables
to do generation once it's trained.

Note that this requires scipy and Pillow are intalled in order to save
the generated images as pngs.